The new boss and "guide". Cheat codes.
Added the first boss (King of slugs). Added cheat codes to speed up game testing. A slight change in the code (transfer of improvement of skills in the procedure). I decided to add a "guide" to facilitate the learning of beginners (so far only the embryo of the "guide"). The plans: to add the ability to anger the "guide" (this will open a new game mode).